edit models postgres
