+++ .gitignore
@@ -5,3 +5,5 @@ dist
 cert
 app.js
 .DS_Store
+./database/postgresdata
+./database/pgadmin
