+++ src/server/api/index.ts
@@ -1,15 +1,22 @@
 import { Sequelize, SequelizeOptions } from 'sequelize-typescript';
+import Messages from './models/postgres/Messages';
+import Topics from './models/postgres/Topics';
+import Users from './models/postgres/Users';
+import Themes from './models/postgres/Themes';
 
-// TODO (ilya): Update options with real ones
 const sequelizeOptions: SequelizeOptions = {
   host: 'localhost',
   port: 5432,
-  username: 'user',
-  password: 'pass',
-  database: 'dbname',
-
+  username: 'yanrun',
+  password: 'yp21',
+  database: 'yanrun_base',
   dialect: 'postgres',
-  models: ['/models'],
+  models: [
+    Users,
+    Topics,
+    Messages,
+    Themes,
+  ],
 };
 
 const sequelize = new Sequelize(sequelizeOptions);
