+++ src/server/api/models/postgres/Messages.ts
@@ -12,7 +12,7 @@ import {
 } from 'sequelize-typescript';
 import Users from './Users';
 import Topics from './Topics';
-import { Nullable } from '../../../client/types';
+import { Nullable } from '../../../../client/types';
 
 type Reactions = {
   like: number[];
@@ -62,23 +62,27 @@ class Messages extends Model<MessageAttributes> {
   })
   topicId!: number;
 
-  @BelongsTo(() => Messages)
-  @Column({
-    type: DataType.INTEGER,
-    field: 'parent_id',
-  })
+  @BelongsTo(() => Messages, 'parent_id')
   parentId!: Nullable<number>;
+
+  // todo(Nail): remove column
+  // @ForeignKey(() => Messages)
+  // @Column({
+  //   type: DataType.INTEGER,
+  //   field: 'messages_id',
+  // })
+  // parentId!: Nullable<number>;
 
   @AllowNull(false)
   @Column(DataType.JSONB)
   reactions!: Reactions;
 
   @AllowNull(false)
+  @CreatedAt
   @Column({
     type: DataType.DATE,
     field: 'created_at',
   })
-  @CreatedAt
   createdAt!: Date;
 }
