+++ src/server/router/app/index.ts
@@ -1,8 +1,7 @@
 import { Router } from 'express';
-import appRoutes from './app';
+import appRoutes from './appRoutes';
 
 const router: Router = Router();
-
 appRoutes(router);
 
 export default router;
