+++ webpack.config.ts
@@ -1,9 +1,7 @@
 import clientConfig from './webpack/client.config';
 import serverConfig from './webpack/server.config';
-import apiServerConfig from './webpack/api-server.config';
 
 export default [
   clientConfig,
   serverConfig,
-  apiServerConfig,
 ];
